int cast

+++ src/simplex_java/Matrica.java
@@ -576,7 +576,7 @@ public class Matrica {
             for(int i=1; i<m+1; i++)
                 for(int j=1; j<n+1; j++)
                 {
-                    matricainverz.get((matrica.get(i).get(0))).set((matrica.get(0).get(j)-n), matrica.get(i).get(j));
+                    matricainverz.get((matrica.get(i).get(0))).set((matrica.get(0).get(j)-n).intValue(), matrica.get(i).get(j));
                 }
             for(int j=1; j<n+1; j++){
                 matrica.get(0).set(j, (j+n)*1.0);
